remove logic from xml_char to skip printing ' '
Remove the logic from xml_char which skipped processing and printing
single spaces. In some cases it was causing words to run together, as
the spaces between them was not getting printed. For example, this
occurred with the version 2.3 of the Sword Project KJV text, exported
as OSIS by diatheke version 4.5 (but was not seen with version 1.0.1
of the sword ESV text).